app/requirements.txt
python-dotenv
rich
prompt-toolkit
requests